Fix invoking virtual function from constructor in WindowsWindow
when calling virtual functions from a base class in your constructor, even if the class you're doing it is the one that defines/overrides the function, this means that potentially the function call is matched at compile-time but the polymorphic behaviour happens at run-time thus, the function call is not using dynamic binding. Thanks cppcheck!